Only kill existing DynamoDB if it already exists

+++ bin/start_dynamodb.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 set -euo pipefail ; IFS=$'\t\n'
 
-kill $(cat dynamodb.pid) || true
+if [ -f dynamodb.pid ]; then
+	kill $(cat dynamodb.pid) || true
+fi
 java -Djava.library.path=./dynamodb_local/DynamoDBLocal_lib -jar dynamodb_local/DynamoDBLocal.jar -inMemory &
 echo $! > dynamodb.pid
 sleep 1
