+++ bin/stop_dynamodb.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
-kill $(cat dynamodb.pid) || true
-echo '==> local dynamo (stopped)'
+if [ -f dynamodb.pid ]; then
+	kill $(cat dynamodb.pid) || true
+	echo '==> local dynamo (stopped)'
+fi
 exit 0
